+++ src/vdir/item/mod.rs
@@ -1,0 +1,7 @@
+pub mod cli;
+pub mod create;
+pub mod delete;
+pub mod get;
+pub mod input;
+pub mod list;
+pub mod update;