+++ src/vdir/mod.rs
@@ -3,5 +3,6 @@ pub mod cli;
 pub mod client;
 pub mod create;
 pub mod delete;
+pub mod item;
 pub mod list;
 pub mod rename;